requirements.txt
# Google ADK core
google-adk==1.16.0

# Gemini API client
google-generativeai==0.8.5

# Web application / environment loading
Flask==3.1.2
python-dotenv==1.1.1

# Data validation
pydantic==2.12.3
